add SBcuda
